Changes to probability validation and mutation
First, for the unbiased_transmission function I've added validations for initial frequency and mutation probabilities so that they never return a negative or >1 values, and if it does, I get a warning and provide a default value (lines 18-27). I'm not 100% if this is the right solution, because it might be too arbitrary (?) and we're sort of interrupting the "natural course" of the simulation.
Secondly, I've modified the conditionals of mutation from A to B and B to A to one single line using ifelse. So if the variant mutates (mutation = TRUE), if A = TRUE it will flip to B, and if A = FALSE (i.e. B) A will be returned (lines 43 - 50).